room: Make Message as a intermediate model b/w Chat and Member.
We are trying to build a many-to-many relation b/w model Chat
and Member.For this, the Message model acts a intermediate model
connecting those two. The message created during the creation of
a new chat is unique indentifier which will used to fetch chats with
a particular member uuid.