server: Bulk subscribe and unsubscribe to rooms.
Instead of subscribing/unsubscribing to pub/sub channel
only on joining/leaving rooms we bulk subscribe/unsubcribe
the channels when the ws singeton is created.

Because of this we recieve the chat notifications
if even we are not inside the room -- this will help us
in updating the order of chats in realtime based on latest
messages.