dm: Add api route to sweep new direct messages.
We use recoil to cache message history. So, when a new message is
sent by a user, the chat is left and returned to, the new message is
no longer visible because the recoil shows the old cached value
and react state containing broadcasted message is cleared. Hence, we
we create a sweeper route to fetch new messages (ie. messages with id
greater than the most recent message in recoil cache.

This is a more efficient approach since we don't fetch entire message
history on every mount.